+++ packages/animated_collection/lib/src/widgets/animated_visibility.dart
@@ -46,6 +46,7 @@ class AnimatedVisibility extends StatelessWidget {
       axisAlignment: axisAlignment,
       curve: curve,
       crossAxisSizeFactor: 1,
+      duration: duration,
       child: visible ? child : const SizedBox.shrink(),
     );
   }