+++ packages/animated_collection/test/widgets/animated_visibility_test.dart
@@ -86,4 +86,53 @@ void main() {
       expect(find.byKey(key), findsNothing);
     },
   );
+
+  testWidgets(
+    'It should use the 1s duration',
+    (tester) async {
+      const key = Key('key');
+      const duration = Duration(seconds: 1);
+      Future<void> pumpWidget(bool visible) async {
+        final widget = Directionality(
+          textDirection: TextDirection.ltr,
+          child: AnimatedVisibility(
+            visible: visible,
+            duration: duration,
+            curve: Curves.linear,
+            child: const ColoredBox(
+              color: Colors.red,
+              child: SizedBox.square(
+                dimension: 50,
+                key: key,
+              ),
+            ),
+          ),
+        );
+        await tester.pumpWidget(widget);
+      }
+
+      await pumpWidget(true);
+
+      final actualPositionedBox = tester.renderObject<RenderPositionedBox>(
+        find.byType(Align),
+      );
+      expect(find.byKey(key), findsOneWidget);
+      expect(actualPositionedBox.heightFactor, 1);
+      expect(actualPositionedBox.widthFactor, 1);
+
+      await pumpWidget(false);
+      await tester.pump(duration * 0.5);
+      expect(find.byKey(key), findsOneWidget);
+      expect(actualPositionedBox.heightFactor, 0.5);
+      expect(actualPositionedBox.widthFactor, 1);
+
+      await tester.pump(duration * 0.5);
+      expect(find.byKey(key), findsOneWidget);
+      expect(actualPositionedBox.heightFactor, 0.0);
+      expect(actualPositionedBox.widthFactor, 1);
+
+      await tester.pumpAndSettle();
+      expect(find.byKey(key), findsNothing);
+    },
+  );
 }